refactor(pages): extract shared nav data-prep helper, eliminate header/sidebar duplication
- Extract `remapVersionedHrefs`, `buildRootMenuItems`, `buildLocaleLinksForNav`,
  `buildSidebarNodes`, and `getThemeDefaultMode` from `_header-with-defaults.tsx`
  and `_sidebar-with-defaults.tsx` into a new shared `pages/lib/_nav-data-prep.ts`
- Both host files and their template mirrors under
  `packages/create-zudo-doc/templates/base/pages/lib/` now import from the shared
  helper; template files are byte-identical to host files (drift check enforces this)
